refactor: use artifact
