+++ .changeset/quick-shoes-lie.md
@@ -2,4 +2,4 @@
 "finagent": patch
 ---
 
-Add /mode command to switch between dashboard and notebook generation 
+Add /mode command to switch between dashboard and notebook generation
